Make compatible with Kaminari without_count
Kaminari's `without_count` stops Kaminari from doing a full count on the scope.

It overrides the `load` method to load in an additional record. If the returned records is > than the user requested limit, then we know there is another page.

`fast_page` wasn't compatible with this because we force the records to be loaded. This PR updates `fast_page` to use the same technique and also sets a `@_has_next` ivar
that kaminari makes use of to skip executing the full count.